gmex-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gmex-sdk"
version = "0.1.0"
description = "Gmail Extraction SDK"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "email-archive",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "pyyaml"
]

[tool.setuptools.packages.find]
where = ["src"]
